tendor aggrigator/agents/etenders.py
"""
Scraping agent for eTenders portal (https://etenders.gov.in)
"""
import requests
from bs4 import BeautifulSoup
import time

def scrape_etenders():
    """
    Scrape tenders from eTenders portal.
    Returns a list of tender dictionaries.
    """
    tenders = []
    
    try:
        # Note: This is a simplified example. The actual implementation would need
        # to handle pagination, authentication, and other complexities.
        url = "https://etenders.gov.in/eprocure/app"
        
        # For demonstration purposes, we'll return sample data
        # In a real implementation, you would scrape the actual website
        
        # Sample tender data
        sample_tenders = [
            {
                "tender_id": "ET-2025-001",
                "organization": "Ministry of Electronics and Information Technology",
                "category": "IT Services",
                "location": "New Delhi",
                "value": "₹25,00,000",
                "deadline": "2025-10-15",
                "description": "Supply and installation of servers and networking equipment",
                "link": "https://etenders.gov.in/eprocure/app?page=FrontEndTenderDetails&service=page&isOpen=1&tenderId=12345"
            },
            {
                "tender_id": "ET-2025-002",
                "organization": "National Highways Authority of India",
                "category": "Construction",
                "location": "Mumbai",
                "value": "₹50,00,00,000",
                "deadline": "2025-11-20",
                "description": "Construction of highway bridge",
                "link": "https://etenders.gov.in/eprocure/app?page=FrontEndTenderDetails&service=page&isOpen=1&tenderId=12346"
            },
            {
                "tender_id": "ET-2025-003",
                "organization": "Indian Railways",
                "category": "Maintenance",
                "location": "Kolkata",
                "value": "₹15,00,000",
                "deadline": "2025-09-30",
                "description": "Annual maintenance of railway signaling equipment",
                "link": "https://etenders.gov.in/eprocure/app?page=FrontEndTenderDetails&service=page&isOpen=1&tenderId=12347"
            }
        ]
        
        tenders.extend(sample_tenders)
        
    except Exception as e:
        print(f"Error scraping eTenders: {e}")
    
    return tenders