tendor aggrigator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
spacy==3.7.2
pymongo==4.6.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytest==7.4.3
streamlit==1.28.0
pandas==2.1.3
openpyxl==3.1.2